Update football.js
football() now assumes the next game will be Sep. 1 of next year if there are no games in the schedule that are in the future. This resolves not getting into the "YES" condition and also handles the condition whereby the current season has ended and the next season's schedule is not available.